+++ sendinblue/target.php
@@ -78,6 +78,11 @@ $search_socname=GETPOST('search_socname', 'none');
 $modulesdir = dolGetModulesDirs('/mailings');
 
 $object = new Mailing($db);
+// Gestion changements v13
+// Gestion de la rétrocompatibilité
+$title = $object->title;
+if (empty ($title)) $title = $object->titre;
+
 $sendinblue= new DolSendinBlue($db);
 $formsendinblue = new FormSendinBlue($db);
 
@@ -247,7 +252,7 @@ if ($object->fetch($id) >= 0)
 	print $form->showrefnav($object,'id', $linkback);
 	print '</td></tr>';
 
-	print '<tr><td width="25%">'.$langs->trans("MailTitle").'</td><td colspan="3">'.$object->titre.'</td></tr>';
+	print '<tr><td width="25%">'.$langs->trans("MailTitle").'</td><td colspan="3">'.$title.'</td></tr>';
 
 	print '<tr><td width="25%">'.$langs->trans("MailFrom").'</td><td colspan="3">'.dol_print_email($object->email_from,0,0,0,0,1).'</td></tr>';
 
@@ -300,6 +305,8 @@ if ($object->fetch($id) >= 0)
 		clearstatcache();
 
 		$var=true;
+        $urlToken = '';
+        if (function_exists('newToken')) $urlToken = "&token=".newToken();
 
 		foreach ($modulesdir as $dir)
 		{
@@ -357,7 +364,7 @@ if ($object->fetch($id) >= 0)
 
 					if ($allowaddtarget)
 					{
-						print '<form name="'.$modulename.'" action="'.$_SERVER['PHP_SELF'].'?action=add&id='.$object->id.'&module='.$modulename.'" method="POST" enctype="multipart/form-data">';
+						print '<form name="'.$modulename.'" action="'.$_SERVER['PHP_SELF'].'?action=add'.$urlToken.'&id='.$object->id.'&module='.$modulename.'" method="POST" enctype="multipart/form-data">';
 						print '<input type="hidden" name="token" value="'.$_SESSION['newtoken'].'">';
 					}
 
@@ -626,7 +633,7 @@ if ($object->fetch($id) >= 0)
 					print '<td align="center">&nbsp;</td>';
 					print '<td align="right" class="nowrap">'.$langs->trans("MailingStatusNotSent");
 					if ($user->rights->mailing->creer && $allowaddtarget) {
-						print '<a href="'.$_SERVER['PHP_SELF'].'?action=delete&rowid='.$obj->rowid.$param.'">'.img_delete($langs->trans("RemoveRecipient"));
+						print '<a href="'.$_SERVER['PHP_SELF'].'?action=delete'.$urlToken.'&rowid='.$obj->rowid.$param.'">'.img_delete($langs->trans("RemoveRecipient"));
 					}
 					print '</td>';
 				}