+++ ChangeLog.md
@@ -1,6 +1,18 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Version 1.1.5 [ 2021-03-04 ]
+
+### Added
+
+### Changed
+
+### Fix
+- Fix warning on list_open.php & list_click.php
+- Fix Compatibility V13 action links
+- Fix Compatibility V13 titre properties changed to title
+- Fix Compatibility V13 contactid properties changed to contact_id
+
 ## Version 1.1.1 [ 2020-12-16 ]
 
 ### Added